Replace CustomForm as Form and rename file acc to convention
