+++ app/components/booking/form.tsx
@@ -16,7 +16,7 @@ import { isFormProcessing } from "~/utils/form";
 import { tw } from "~/utils/tw";
 import { ActionsDropdown } from "./actions-dropdown";
 import CustodianUserSelect from "../custody/custodian-user-select";
-import { CustomForm as Form } from "../CustomForm";
+import { Form } from "../custom-form";
 import FormRow from "../forms/form-row";
 import Input from "../forms/input";
 import { AbsolutePositionedHeaderActions } from "../layout/header/absolute-positioned-header-actions";
